requirements.txt
aniso8601==7.0.0
APScheduler==3.6.3
argon2-cffi==20.1.0
asgiref==3.4.1
autopep8==1.5.7
bottle==0.12.19
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
django-telegrambot==1.0.1
djangorestframework==3.12.4
environ==1.0
graphene==2.1.9
graphene-django==2.15.0
graphene-django-optimizer==0.8.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.1.0
idna==3.2
promise==2.3
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0
